@@ -9,48 +9,51 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20120511023550) do
+ActiveRecord::Schema.define(version: 20120102201205) do
 
-  create_table "codes", :force => true do |t|
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "codes", force: :cascade do |t|
     t.integer  "location_id"
     t.string   "code"
     t.string   "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",  null: false
+    t.datetime "updated_at",  null: false
   end
 
-  create_table "locations", :force => true do |t|
+  create_table "locations", force: :cascade do |t|
     t.string   "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "users", :force => true do |t|
-    t.string   "login",                                   :null => false
-    t.string   "password",                                :null => false
-    t.string   "role",              :default => "user",   :null => false
-    t.string   "name",                                    :null => false
-    t.string   "email",                                   :null => false
-    t.string   "address",                                 :null => false
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.string   "rate_type",         :default => "hourly", :null => false
-    t.integer  "rate_amount_cents", :default => 0,        :null => false
-    t.string   "invoice_recipient"
+  create_table "users", force: :cascade do |t|
+    t.string   "login",                                null: false
+    t.string   "password",                             null: false
+    t.string   "role",              default: "user",   null: false
+    t.string   "name",                                 null: false
+    t.string   "email",                                null: false
+    t.string   "address",                              null: false
+    t.string   "invoice_recipient", default: "N/A"
+    t.datetime "created_at",                           null: false
+    t.datetime "updated_at",                           null: false
+    t.string   "rate_type",         default: "hourly", null: false
+    t.integer  "rate_amount_cents", default: 0,        null: false
   end
 
-  create_table "work_entries", :force => true do |t|
+  create_table "work_entries", force: :cascade do |t|
     t.integer  "code_id"
     t.date     "worked_at"
     t.integer  "minutes"
     t.integer  "rate_cents"
     t.integer  "fee_cents"
     t.text     "note"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.integer  "user_id",    :default => 2, :null => false
+    t.datetime "created_at",             null: false
+    t.datetime "updated_at",             null: false
+    t.integer  "user_id",    default: 2, null: false
   end
 
 end
