@@ -44,7 +44,7 @@ class LocationsController < ApplicationController
   # POST /locations
   # POST /locations.json
   def create
-    @location = Location.new(params[:location])
+    @location = Location.new(params.permit(:location))
 
     respond_to do |format|
       if @location.save
@@ -63,7 +63,7 @@ class LocationsController < ApplicationController
     @location = Location.find(params[:id])
 
     respond_to do |format|
-      if @location.update_attributes(params[:location])
+      if @location.update_attributes(params.permit(:location))
         format.html { redirect_to @location, notice: 'Location was successfully updated.' }
         format.json { head :ok }
       else
