@@ -1,6 +1,6 @@
 class Code < ActiveRecord::Base
   belongs_to :location
-  has_many :work_entries, :order => 'worked_at, created_at'
+  has_many :work_entries, -> { order('worked_at, created_at') }
 
   scope :display_order, lambda { includes(:location).order('locations.name', 'codes.name') }
 
