@@ -4,12 +4,14 @@ Timelord::Application.configure do
   # Code is not reloaded between requests
   config.cache_classes = true
 
+  config.eager_load = false
+
   # Full error reports are disabled and caching is turned on
   config.consider_all_requests_local       = false
   config.action_controller.perform_caching = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = false
+  config.serve_static_files = false
 
   # Compress JavaScripts and CSS
   config.assets.compress = true
