@@ -1,9 +1,9 @@
 Timelord::Application.routes.draw do
 
-  match 'account/login', :controller => 'account', :action => 'login'
-  match 'account/logout', :controller => 'account', :action => 'logout'
-  match 'account/:id/edit', :controller => 'account', :action => 'edit'
-  match 'account/1', :method => :put, :controller => 'account', :action => 'update'
+  get 'account/login', :controller => 'account', :action => 'login'
+  get 'account/logout', :controller => 'account', :action => 'logout'
+  get 'account/:id/edit', :controller => 'account', :action => 'edit'
+  put 'account/1', :controller => 'account', :action => 'update'
 
   resources :users
 
@@ -13,69 +13,12 @@ Timelord::Application.routes.draw do
 
   resources :locations
 
-  match '/', :controller => 'work_entries', :action => 'new'
-  match 'nav', :controller => 'nav', :action => 'index'
-  match 'nav/:action', :controller => 'nav'
+  get '/', :controller => 'work_entries', :action => 'new'
+  get 'nav', :controller => 'nav', :action => 'index'
+  get 'nav/:action', :controller => 'nav'
 
-  match 'locations/:id/codes(.:format)', :controller => 'locations', :action => 'codes'
+  get 'locations/:id/codes(.:format)', :controller => 'locations', :action => 'codes'
 
-  match 'invoice', :controller => 'invoice', :action => 'index'
-  match 'invoice/preview(.:format)', :controller => 'invoice', :action => 'preview'
-
-  # The priority is based upon order of creation:
-  # first created -> highest priority.
-
-  # Sample of regular route:
-  #   match 'products/:id' => 'catalog#view'
-  # Keep in mind you can assign values other than :controller and :action
-
-  # Sample of named route:
-  #   match 'products/:id/purchase' => 'catalog#purchase', :as => :purchase
-  # This route can be invoked with purchase_url(:id => product.id)
-
-  # Sample resource route (maps HTTP verbs to controller actions automatically):
-  #   resources :products
-
-  # Sample resource route with options:
-  #   resources :products do
-  #     member do
-  #       get 'short'
-  #       post 'toggle'
-  #     end
-  #
-  #     collection do
-  #       get 'sold'
-  #     end
-  #   end
-
-  # Sample resource route with sub-resources:
-  #   resources :products do
-  #     resources :comments, :sales
-  #     resource :seller
-  #   end
-
-  # Sample resource route with more complex sub-resources
-  #   resources :products do
-  #     resources :comments
-  #     resources :sales do
-  #       get 'recent', :on => :collection
-  #     end
-  #   end
-
-  # Sample resource route within a namespace:
-  #   namespace :admin do
-  #     # Directs /admin/products/* to Admin::ProductsController
-  #     # (app/controllers/admin/products_controller.rb)
-  #     resources :products
-  #   end
-
-  # You can have the root of your site routed with "root"
-  # just remember to delete public/index.html.
-  root :to => 'work_entries#new'
-
-  # See how all your routes lay out with "rake routes"
-
-  # This is a legacy wild controller route that's not recommended for RESTful applications.
-  # Note: This route will make all actions in every controller accessible via GET requests.
-  # match ':controller(/:action(/:id(.:format)))'
+  get 'invoice', :controller => 'invoice', :action => 'index'
+  get 'invoice/preview(.:format)', :controller => 'invoice', :action => 'preview'
 end
