@@ -2,8 +2,7 @@ class CreateLocations < ActiveRecord::Migration
   def change
     create_table :locations do |t|
       t.string :name
-
-      t.timestamps
+      t.timestamps null: false
     end
   end
 end
